Move existing tests to unit
